# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API prueba previa junior PHP",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/user": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Mostrar usuarios",
                "operationId": "971db9bc974c0e607288cfe7ecc10096",
                "responses": {
                    "200": {
                        "description": "Mostrar todos los usuarios.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Crear usuario",
                "operationId": "089999749fbdec98922de2f74ba8d540",
                "requestBody": {
                    "request": "UserStoreRequest",
                    "description": "Contiene los datos del usuario a añadir",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {},
                            "example": {
                                "user": {
                                    "email": "test@gmail.com",
                                    "password": "12341234",
                                    "name": "prueba"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuario creado de manera exitosa",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "default": {
                        "description": "Ha ocurrido un error.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user/{id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Mostrar un usuario",
                "operationId": "d8f696420bea235f4916124c9ba1c507",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "el id del usuario a mostrar",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mostrar un usuarios.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Borrar un usuario",
                "operationId": "f8d7aea516ee464d0afee90f3a1a1722",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "el id del usuario a eliminar",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mostrar todos los usuarios.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Autenticacion"
                ],
                "summary": "Login",
                "operationId": "0fc9610184083ac6dddab2c63139225a",
                "requestBody": {
                    "request": "LoginRequest",
                    "description": "Contiene el email y la contraseña a loggear",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {},
                            "example": {
                                "user": {
                                    "email": "test@gmail.com",
                                    "password": "12341234"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login exitoso",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "default": {
                        "description": "Ha ocurrido un error.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/logout": {
            "get": {
                "tags": [
                    "Autenticacion"
                ],
                "summary": "Cerrar sesión",
                "operationId": "8992dd4fbd21ecd6bd79f84228dc98cf",
                "responses": {
                    "200": {
                        "description": "Cerrar sesión",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}